R code
The list of potential counterfactual units will be named as the donor pool in which to observe the evolution of the outcome before and after the treatment. As main concerns of the data selection, for this group of prospective comparison units, data accessibility, and where Sri Lanka introduced the liberalization policy, in which all the nations for which there are statistics that were unaffected by the treatment event. The potential control nations will be summarized as follows except the country id 1 which is LKA as Sri Lanka.

Additionally, the outcome predictors for these counterfactual units will be required. We are looking for factors that will make it possible to forecast how the synthetic control group will respond to the treatment in 1977. Moreover, the synthetic control group with average values for these variables are comparable to those of the treated unit.